pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "open-match-lca"
version = "0.1.0"
description = "Open-data-first LCA matching and EPA factor prediction research scaffold"
readme = "README.md"
requires-python = ">=3.11,<3.12"
license = {text = "MIT"}
authors = [
  {name = "Open Match LCA Contributors"}
]
dependencies = [
  "numpy>=1.26",
  "pandas>=2.2",
  "pyarrow>=15.0",
  "PyYAML>=6.0",
  "requests>=2.31",
  "scikit-learn>=1.4",
  "scipy>=1.12",
  "rank-bm25>=0.2.2",
  "matplotlib>=3.8",
  "seaborn>=0.13",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
]
full = [
  "sentence-transformers>=3.0",
  "lightgbm>=4.3",
  "xgboost>=2.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
